+++ lib/bindable.dart
@@ -1,8 +1,11 @@
-///
-/// yichen <d.unicreators@gmail.com>
-///
+// Copyright (c) 2019 yichen <d.unicreators@gmail.com>. All rights reserved.
+// Use of this source code is governed by a MIT license that can be
+// found in the LICENSE file.
 
 part of './mvvm.dart';
+
+typedef PropertyValueChanged<TValue> = void Function(
+    TValue value, Object propertyKey);
 
 /// BindableObject
 ///
@@ -17,7 +20,6 @@ abstract class BindableObject {
 
   ///
   /// 注册一个属性
-  @protected
   BindableProperty<TValue> registryProperty<TValue>(
       BindableProperty<TValue> property) {
     assert(property != null);
@@ -57,9 +59,21 @@ abstract class BindableObject {
 
   ///
   /// 设置指定 [propertyKey] 对应的属性值
+  ///
+  /// [value] 指定属性值
+  /// [valueCheck] 指定是否对值进行检查,
+  ///   当其值为 `true` 时, 多次设置相同值将不会触发值变更通知
+  ///   默认为 `false`
   @protected
-  void setValue<TValue>(Object propertyKey, TValue value) =>
+  void setValue<TValue>(Object propertyKey, TValue value,
+      {bool valueCheck = false}) {
+    if (valueCheck) {
+      var oldValue = getValue<TValue>(propertyKey);
+      if (oldValue != value) getProperty<TValue>(propertyKey)?.value = value;
+    } else {
       getProperty<TValue>(propertyKey)?.value = value;
+    }
+  }
 
   ///
   /// 获取指定 [propertyKey] 对应的属性 [ValueListenable]
@@ -81,21 +95,54 @@ abstract class BindableObject {
   Iterable<ValueListenable<TValue>> getValueListenables<TValue>(
           Iterable<Object> propertyKeys) =>
       propertyKeys.map(getValueListenable);
+
+  /// dispose
+  @protected
+  @mustCallSuper
+  void dispose() {
+    if (!_properties.isNotEmpty) return;
+    for (var prop in _properties.values) {
+      prop.dispose();
+    }
+  }
+}
+
+class _BindableProperty<TValue> extends BindableProperty<TValue> {
+  _BindableProperty(Object key, ValueNotifier<TValue> valueNotifier,
+      {PropertyValueChanged<TValue> valueChanged})
+      : super(key, valueNotifier, valueChanged: valueChanged);
 }
 
 /// BindableProperty
 ///
 abstract class BindableProperty<TValue> {
+  /// BindableProperty.create
+  static BindableProperty<TValue> create<TValue>(
+          Object key, ValueNotifier<TValue> valueNotifier,
+          {PropertyValueChanged<TValue> valueChanged}) =>
+      _BindableProperty(key, valueNotifier, valueChanged: valueChanged);
+
   /// key
   final Object key;
   final ValueNotifier<TValue> _valueNotifier;
+  VoidCallback _listener;
 
   /// BindableProperty
-  BindableProperty(this.key, this._valueNotifier
-      /* , {TValue initial} */)
-      : assert(key != null && _valueNotifier != null);
+  BindableProperty(this.key, this._valueNotifier,
+      {PropertyValueChanged<TValue> valueChanged})
+      : assert(key != null && _valueNotifier != null) {
+    if (valueChanged != null) {
+      var _listener = () => valueChanged(value, key);
+      _valueNotifier.addListener(_listener);
+    }
+  }
 
-  //ValueNotifier<TValue> get valueNotifier => _valueNotifier;
+  /// dispose
+  @protected
+  @mustCallSuper
+  void dispose() {
+    if (_listener != null) _valueNotifier.removeListener(_listener);
+  }
 
   ///
   /// 获取值