+++ lib/property_adaptive.dart
@@ -1,6 +1,6 @@
-///
-/// yichen <d.unicreators@gmail.com>
-///
+// Copyright (c) 2019 yichen <d.unicreators@gmail.com>. All rights reserved.
+// Use of this source code is governed by a MIT license that can be
+// found in the LICENSE file.
 
 part of './mvvm.dart';
 
@@ -14,12 +14,14 @@ class AdaptiveViewModelProperty<TValue, TAdaptee extends Listenable>
       TAdaptee adaptee,
       TValue Function(TAdaptee) getAdapteeValue,
       void Function(TAdaptee, TValue) setAdapteeValue,
-      {TValue initial})
+      {PropertyValueChanged<TValue> valueChanged,
+      TValue initial})
       : super(
             key,
             ValueNotifierAdapter<TValue, TAdaptee>(
                 adaptee, getAdapteeValue, setAdapteeValue,
-                initial: initial));
+                initial: initial),
+            valueChanged: valueChanged);
 }
 
 mixin AdaptiveViewModelMixin on _ViewModelBase {
@@ -30,6 +32,8 @@ mixin AdaptiveViewModelMixin on _ViewModelBase {
   /// [adaptee] 被适配者实例，适配者必须继承自 [Listenable]
   /// [getAdapteeValue] 指定从被适配者获取值的方法
   /// [setAdapteeValue] 指定设置被适配者值的方法
+  /// [valueChanged] 指定属性值变更后的回调方法
+  /// [initial] 指定初始值
   ///
   BindableProperty<TValue>
       propertyAdaptive<TValue, TAdaptee extends Listenable>(
@@ -37,8 +41,9 @@ mixin AdaptiveViewModelMixin on _ViewModelBase {
               TAdaptee adaptee,
               TValue Function(TAdaptee) getAdapteeValue,
               void Function(TAdaptee, TValue) setAdapteeValue,
-              {TValue initial}) =>
+              {PropertyValueChanged<TValue> valueChanged,
+              TValue initial}) =>
           registryProperty(AdaptiveViewModelProperty<TValue, TAdaptee>(
               propertyKey, adaptee, getAdapteeValue, setAdapteeValue,
-              initial: initial));
+              valueChanged: valueChanged, initial: initial));
 }